Move internal stuff in internal, fix some styling issues
