+++ internal/caddy/tls.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"github.com/caddyserver/caddy/v2/modules/caddyhttp"
 	"github.com/caddyserver/caddy/v2/modules/caddytls"
-	"github.com/caddyserver/ingress/pkg/controller"
+	"github.com/caddyserver/ingress/internal/controller"
 )
