+++ internal/controller/action_status.go
@@ -1,7 +1,7 @@
 package controller
 
 import (
-	"github.com/caddyserver/ingress/pkg/k8s"
+	"github.com/caddyserver/ingress/internal/k8s"
 	"go.uber.org/zap"
 	"gopkg.in/go-playground/pool.v3"
 	apiv1 "k8s.io/api/core/v1"
