+++ internal/controller/store.go
@@ -1,7 +1,7 @@
 package controller
 
 import (
-	"github.com/caddyserver/ingress/pkg/k8s"
+	"github.com/caddyserver/ingress/internal/k8s"
 	"k8s.io/api/networking/v1beta1"
 )
